Fix TSE review
