--- src/actions/progressAction.js
@@ -1,34 +1,0 @@
-import axios from 'axios';
-import { PROGRESS_CALCULATION, STUDIES_ERRORS } from '../constants/actionTypes';
-
-// const defaultURL = 'http://localhost:3001/api/v1'; // development
-
-const defaultURL = 'https://pure-stream-80472.herokuapp.com/api/v1'; // production
-
-const apiConfig = {
-  headers: {
-    'Content-Type': 'application/json',
-    Accept: 'application/json',
-    Authorization: `Bearer ${localStorage.getItem('token')}`,
-    'Access-Control-Allow-Origin': '*',
-    'Access-Control-Allow-Headers': '*',
-  },
-};
-// eslint-disable-next-line consistent-return
-const progressCal = () => async (dispatch) => {
-  try {
-    const response = await axios.get(`${defaultURL}/levels/progress`, apiConfig);
-    dispatch({
-      type: PROGRESS_CALCULATION,
-      payload: response.data.progress,
-    });
-    return response;
-  } catch (error) {
-    dispatch({
-      type: STUDIES_ERRORS,
-      payload: error,
-    });
-  }
-};
-
-export default progressCal;
