+++ src/store.js
@@ -2,5 +2,8 @@ import { createStore, applyMiddleware, compose } from 'redux';
 import thunk from 'redux-thunk';
 import rootReducer from './reducers';
 
-const store = createStore(rootReducer, {}, compose(applyMiddleware(thunk)));
+const initialState = {};
+
+const store = createStore(rootReducer, initialState, compose(applyMiddleware(thunk)));
+
 export default store;
